requirements.txt
fastapi
uvicorn
pydantic[email]
supabase
python-dotenv
